Move the database creation into a standalone script
