+++ pages/api/index.ts
@@ -1,5 +1,5 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
-import { initDatabase } from '../../lib/init';
+import { initDatabase } from '../../scripts/init';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
   await initDatabase(); // Ensure DB and tables exist
